CTFd/plugins/ctfd-owl/requirements.txt
docker-compose
PyYAML >= 5
